requirements.txt
azure-functions==1.23.0
pyodbc==5.2.0
azure-identity==1.21.0
azure-keyvault-secrets==4.9.0
